rename MathLib/LinAlg/LinearEquation to MathLib/LinAlg/SystemOfLinearEquations
